backend/requirements.txt
fastapi
uvicorn
ruff
Pillow
puremagic
